Fixed University Entity, Repository, Service

+++ src/main/java/com/example/csws/repository/university/JpaUniversityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface JpaUniversityRepository extends JpaRepository<University, Integer>, UniversityRepository {
+public interface JpaUniversityRepository extends JpaRepository<University, Long>, UniversityRepository {
 
     public List<University> findAll();
 }
